// angular-data-dashboard/src/app/google-chart/model/data.model.ts
export interface Root extends Array<Root> {
  id: number
  end_year: number
  citylng: string
  citylat: string
  intensity: number
  sector: string
  topic: string
  insight: string
  swot: string
  url: string
  region: string
  start_year: number
  impact: number
  added: string
  published: string
  city: string
  country: string
  relavance: number
  pestle: string
  source: string
  title: string
  likelihood: number
}